requirements.txt
langchain
langsmith
openai
youtube-transcript-api
sentence-transformers
python-dotenv
streamlit
langgraph
langchain_openai
langchain_community
chromadb
requests
